Finalize implementations

+++ cpp/main.cpp
@@ -1,4 +1,5 @@
 #include <chrono>
+#include <cstdint>
 #include <functional>
 #include <iostream>
 #include <string>
@@ -8,8 +9,8 @@ void print_header() {
   std::cout << "language, benchmark, run_id, time" << std::endl;
 }
 
-void measure(std::function<void(void)> act, std::string label, int num_iters) {
-  for (int it_num = 0; it_num < num_iters; it_num++) {
+void measure(std::function<void(void)> act, std::string label, int64_t num_iters) {
+  for (int64_t it_num = 0; it_num < num_iters; it_num++) {
     auto x1 = std::chrono::steady_clock::now();
     act();
     auto x2 = std::chrono::steady_clock::now();
@@ -19,9 +20,9 @@ void measure(std::function<void(void)> act, std::string label, int num_iters) {
   }
 }
 
-int sum(int n) {
-  int acc = 0;
-  int i = 1;
+int64_t sum(int64_t n) {
+  int64_t acc = 0;
+  int64_t i = 1;
   while (i <= n) {
     acc += i;
     i++;
@@ -29,16 +30,16 @@ int sum(int n) {
   return acc;
 }
 
-int* alloc_vector(int n) {
-  int *res = new int[n];
-  for (int i = 0; i < n; ++i) {
+int64_t* alloc_vector(int64_t n) {
+  int64_t *res = new int64_t[n];
+  for (int64_t i = 0; i < n; ++i) {
     res[i] = i;
   }
   return res;
 }
 
-int sum_vector(const int *vec, size_t size) {
-  int acc = 0;
+int64_t sum_vector(const int64_t *vec, size_t size) {
+  int64_t acc = 0;
   for (size_t i = 0; i < size; ++i) {
     acc += vec[i];
   }
@@ -46,20 +47,20 @@ int sum_vector(const int *vec, size_t size) {
 }
 
 struct Cons {
-  int head;
+  int64_t head;
   Cons *tail;
 };
 
-Cons* alloc_list(int n) {
+Cons* alloc_list(int64_t n) {
   Cons *res = nullptr;
-  for (int i = n; i >= 0; --i) {
+  for (int64_t i = n; i >= 0; --i) {
     res = new Cons{i, res};
   }
   return res;
 }
 
-int sum_list(Cons *list) {
-  int acc = 0;
+int64_t sum_list(Cons *list) {
+  int64_t acc = 0;
   while (list != nullptr) {
     acc += list->head;
     list = list->tail;
@@ -68,32 +69,32 @@ int sum_list(Cons *list) {
 
 struct Tree {
   Tree *left;
-  int elem;
+  int64_t elem;
   Tree *right;
 };
 
-Tree* allocate_helper(int& ix, int remaining_depth) {
+Tree* allocate_helper(int64_t& ix, int64_t remaining_depth) {
   if (remaining_depth == 0) {
     return nullptr;
   } else {
     auto l = allocate_helper(ix, remaining_depth - 1);
-    int e = ix++;
+    int64_t e = ix++;
     auto r = allocate_helper(ix, remaining_depth - 1);
     return new Tree{l, e, r};
   }
 }
 
-Tree* alloc_full_tree(int depth) {
-  int ix = 0;
+Tree* alloc_full_tree(int64_t depth) {
+  int64_t ix = 0;
   return allocate_helper(ix, depth);
 }
 
-int sum_tree(Tree* tree) {
+int64_t sum_tree(Tree* tree) {
   if (tree == nullptr) {
     return 0;
   } else {
-    int l = sum_tree(tree->left);
-    int r = sum_tree(tree->right);
+    int64_t l = sum_tree(tree->left);
+    int64_t r = sum_tree(tree->right);
     return l + tree->elem + r;
   }
 }
@@ -103,9 +104,9 @@ int main() {
   // size = 100000000
   // depth = 25
   // n_iters = 10
-  int size = 100; //TODO FIXME temporary workaround to quickly test
-  int depth = 4;
-  int n_iters = 2;
+  int64_t size = 100; //TODO FIXME temporary workaround to quickly test
+  int64_t depth = 4;
+  int64_t n_iters = 2;
 
   measure([]() { std::this_thread::sleep_for(std::chrono::milliseconds(100)); }, "100ms", 10);
